+++ rust/cedar-policy-mcp-schema-generator/src/lib.rs
@@ -14,17 +14,8 @@
  * limitations under the License.
  */
 
-mod description;
-mod err;
-mod schema_generator;
-
-pub use description::{
-    err::DeserializationError, parser::err::ParseError, Parameters, Property, PropertyType,
-    PropertyTypeDef, ServerDescription, ToolDescription,
-};
-
-pub use err::SchemaGeneratorError;
-pub use schema_generator::{SchemaGenerator, SchemaGeneratorConfig};
+mod generator;
+pub use generator::{SchemaGenerator, SchemaGeneratorConfig, SchemaGeneratorError};
 
 #[cfg(feature = "cli")]
 mod cli;